refactor(domains): move brand-claims to domains/vendors (#232)
