+++ src/domains/vendors/brand-claims.ts
