+++ src/i18n/public-page-copy.ts
@@ -1,6 +1,6 @@
 // Static page content module. See ./README.md for when to use *-copy.ts vs flat keys.
 import { DEFAULT_COMMISSION_RATE } from '@/lib/constants'
-import { BRAND_CLAIMS } from '@/lib/brand-claims'
+import { BRAND_CLAIMS } from '@/domains/vendors/brand-claims'
 import type { Locale } from './locales'
 
 type FaqItem = {
